Утилиты для фигур; Орфография Car
Утилиты для фигур; Орфография Car

+++ lab04/Shapes/CCanvas.h
@@ -7,10 +7,10 @@ class CCanvas : public ICanvas
 public:
 	CCanvas();
 	void Display();
-	void DrawLine(CPoint from, CPoint to, unsigned lineColor) override;
-	void FillPolygon(std::vector<CPoint> points, unsigned fillColor) override;
-	void DrawCircle(CPoint center, double radius, unsigned lineColor) override;
-	void FillCircle(CPoint center, double radius, unsigned fillColor) override;
+	void DrawLine(CPoint from, CPoint to, unsigned long lineColor) override;
+	void FillPolygon(std::vector<CPoint> points, unsigned long fillColor) override;
+	void DrawCircle(CPoint center, double radius, unsigned long lineColor) override;
+	void FillCircle(CPoint center, double radius, unsigned long fillColor) override;
 private:
 	sf::RenderWindow m_window;
 };
